+++ change_log.md
@@ -9,3 +9,15 @@ v0.1
     - Admin to view and update data
     - Remaining classes logic from add_attendance()
 
+v0.2
+ - Have all basic functions
+    - AdminScreen: Admin to create user, studnet, attendance
+    - LoginScreen: User auth login with Firebase
+    - StudentSelectionScreen: User to choose a student to view
+    - AttendanceScreen: User to view attendance records
+    - AdminViewUser: Show all user
+    - AdminViewStudent: Show all students
+    - AdminViewAttendance: Show all attendances and delete
+ - Next step:
+    - Push to web
+